requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai==0.8.5
tqdm
protobuf
pydantic
